Updated the Word bank and Hint AGAIN
i didnt know that we were supposed to read the words from provided file. so in this update i made the program read the words in a file and write them to the word array in the program. now the hint is actually a free letter that is given to the user when the hint option is selected.